+++ packages/graphiql/src/components/QueryHistory.tsx
@@ -22,6 +22,7 @@ const MAX_HISTORY_LENGTH = 20;
 const shouldSaveQuery = (
   query?: string,
   variables?: string,
+  headers?: string,
   lastQuerySaved?: QueryStoreItem,
 ) => {
   if (!query) {
@@ -45,7 +46,12 @@ const shouldSaveQuery = (
     if (
       JSON.stringify(variables) === JSON.stringify(lastQuerySaved.variables)
     ) {
-      return false;
+      if (JSON.stringify(headers) === JSON.stringify(lastQuerySaved.headers)) {
+        return false;
+      }
+      if (headers && !lastQuerySaved.headers) {
+        return false;
+      }
     }
     if (variables && !lastQuerySaved.variables) {
       return false;
@@ -57,6 +63,7 @@ const shouldSaveQuery = (
 type QueryHistoryProps = {
   query?: string;
   variables?: string;
+  headers?: string;
   operationName?: string;
   queryID?: number;
   onSelectQuery: HandleSelectQueryFn;
@@ -118,12 +125,21 @@ export class QueryHistorySource extends React.Component<
   updateHistory = (
     query?: string,
     variables?: string,
+    headers?: string,
     operationName?: string,
   ) => {
-    if (shouldSaveQuery(query, variables, this.historyStore.fetchRecent())) {
+    if (
+      shouldSaveQuery(
+        query,
+        variables,
+        headers,
+        this.historyStore.fetchRecent(),
+      )
+    ) {
       this.historyStore.push({
         query,
         variables,
+        headers,
         operationName,
       });
       const historyQueries = this.historyStore.items;
@@ -139,6 +155,7 @@ export class QueryHistorySource extends React.Component<
   toggleFavorite: HandleToggleFavoriteFn = (
     query,
     variables,
+    headers,
     operationName,
     label,
     favorite,
@@ -146,6 +163,7 @@ export class QueryHistorySource extends React.Component<
     const item: QueryStoreItem = {
       query,
       variables,
+      headers,
       operationName,
       label,
     };
@@ -165,6 +183,7 @@ export class QueryHistorySource extends React.Component<
   editLabel: HandleEditLabelFn = (
     query,
     variables,
+    headers,
     operationName,
     label,
     favorite,
@@ -172,6 +191,7 @@ export class QueryHistorySource extends React.Component<
     const item = {
       query,
       variables,
+      headers,
       operationName,
       label,
     };