+++ packages/graphiql/src/types/index.ts
@@ -1,8 +1,6 @@
 /** global monaco */
 // eslint-disable-next-line spaced-comment
 /// <reference path='../../../../node_modules/monaco-editor/monaco.d.ts'/>
-// eslint-disable-next-line spaced-comment
-/// <reference path='../../../../packages/monaco-graphql/src/typings/monaco.d.ts'/>
 
 import { GraphQLType } from 'graphql';