+++ packages/graphql-language-service-interface/src/getDiagnostics.ts
@@ -120,9 +120,9 @@ function annotations(
   const highlightedNodes: Diagnostic[] = [];
   error.nodes.forEach(node => {
     const highlightNode =
-      node.kind !== 'Variable' && 'name' in node
+      node.kind !== 'Variable' && 'name' in node && node.name !== undefined
         ? node.name
-        : 'variable' in node
+        : 'variable' in node && node.variable !== undefined
         ? node.variable
         : node;
     if (highlightNode) {