+++ examples/graphiql-parcel/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.5](https://github.com/graphql/graphiql/compare/graphiql-parcel-example@1.1.4...graphiql-parcel-example@1.1.5) (2020-06-11)
+
+**Note:** Version bump only for package graphiql-parcel-example
+
+## [1.1.4](https://github.com/graphql/graphiql/compare/graphiql-parcel-example@1.1.3...graphiql-parcel-example@1.1.4) (2020-06-04)
+
+**Note:** Version bump only for package graphiql-parcel-example
+
+## [1.1.3](https://github.com/graphql/graphiql/compare/graphiql-parcel-example@1.1.2...graphiql-parcel-example@1.1.3) (2020-06-04)
+
+**Note:** Version bump only for package graphiql-parcel-example
+
+## [1.1.2](https://github.com/graphql/graphiql/compare/graphiql-parcel-example@1.1.1...graphiql-parcel-example@1.1.2) (2020-05-28)
+
+**Note:** Version bump only for package graphiql-parcel-example
+
+## [1.1.1](https://github.com/graphql/graphiql/compare/graphiql-parcel-example@1.1.0...graphiql-parcel-example@1.1.1) (2020-05-19)
+
+**Note:** Version bump only for package graphiql-parcel-example
+
 # 1.1.0 (2020-05-17)
 
 ### Features